requirements.txt
Django==2.1.3
django-extensions==2.1.4
djangorestframework==3.9.0
drf-nested-routers==0.91
gunicorn==19.9.0
pytz==2018.7
six==1.11.0
